refactor(seiri): reorganize constants and improve workflow handling
- Moved `WORKFLOW_CHAIN_LINE` to `postureLines.ts` and updated references accordingly.
- Introduced new constants: `ELECTION_STANDARD_LINE`, `ELECTION_STRICT_LINE`, and `ELECTION_RENDER` in `electionLines.ts`.
- Added `FAILURE_CHAIN_THRESHOLD`, `TRACKED_COMMANDS_CAP`, and `FAILURE_CHAIN_LINE` in `failureChain.ts`.
- Created `workflowChain.ts` and `workflowStateLines.ts` to manage workflow skills and their states.
- Updated imports across various files to reflect the new structure, ensuring consistency and clarity in the codebase.
- Removed obsolete `signals.ts` file to streamline the constants used in the project.

Verified: All tests passed successfully, maintaining functionality and performance.
