+++ plugins/seiri/src/constants/electionLines.ts
@@ -1,0 +1,39 @@
+import { INTERVENTION } from './intervention.js';
+
+/**
+ * The election contract (D7-E, Arm S), restated by both injection
+ * channels: SubagentStart, where a spawn inherits none of the parent's
+ * SessionStart context, and SessionStart itself. It sits beside
+ * `WORKFLOW_CHAIN_LINE` (constants/postureLines.ts) rather than replacing
+ * it — the chain says which workflow follows which, this says a matching
+ * moment is loaded before it is acted on. Neither touches the deployed
+ * rule documents, whose bytes must keep matching templateHash.
+ *
+ * `standard` frames the procedure and names exactly one skill: the
+ * done-claim moment, where the measured standard/strict gap concentrated.
+ * `strict` names every moment's owner outright. Both force election
+ * (load-order) only — adoption (a stated-reason deviation after reading)
+ * stays with the model, which is why both end on "decide after reading" /
+ * "deviations are yours to make, with a stated reason".
+ *
+ * In its own file so only the render path carries these bytes: every
+ * dial consumer imports `constants/intervention.ts`, and the
+ * computed-key map below is the one statement esbuild cannot prove
+ * pure — inlined there, it rode into every hook and the MCP server.
+ */
+export const ELECTION_STANDARD_LINE =
+  'Election: defined workflows govern these moments — a failure appearing, multi-step work starting, review arriving or departing — load the matching seiri workflow before acting. One moment is named: before saying done, fixed, or passing, load `/seiri:verify`. Decide after reading — deviations are yours to make, with a stated reason.';
+
+export const ELECTION_STRICT_LINE =
+  'Election contract: these moments have owners — a failure appears → load `/seiri:trace-cause` · multi-step work begins → `/seiri:write-plan` · a plan exists → `/seiri:execute` · before implementing → `/seiri:implement` · before saying done/fixed/passing → `/seiri:verify` · requesting review → `/seiri:request-review` · feedback arrives → `/seiri:receive-review`. Load first, decide after reading — deviations are yours to make, with a stated reason. Matching without loading is a skipped election, not a judgment.';
+
+/**
+ * Dial → election text. `advisory` has no entry — the lookup misses and
+ * every caller renders silence, keeping that dial position exactly as the
+ * dispatch measurements were taken against. A missing entry, not a branch
+ * in each caller, is what makes the opt-out hold across channels.
+ */
+export const ELECTION_RENDER = {
+  [INTERVENTION.STANDARD]: ELECTION_STANDARD_LINE,
+  [INTERVENTION.STRICT]: ELECTION_STRICT_LINE,
+} as const;
